refactor(agent): extract history and context helpers
